LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-text/i18n/zh/css-text-line-break-zh-in-loose-expected.txt


FAIL 2024  ONE DOT LEADER may appear at line start if zh and loose assert_approx_equals: expected 49.02726745605469 +/- 1 but got 79.35726928710938
FAIL 2025  TWO DOT LEADER may appear at line start if zh and loose assert_approx_equals: expected 69.66000366210938 +/- 1 but got 99.98999786376953
FAIL 2026  HORIZONTAL ELLIPSIS may appear at line start if zh and loose assert_approx_equals: expected 69.66000366210938 +/- 1 but got 99.98999786376953
FAIL 22EF  MIDLINE HORIZONTAL ELLIPSIS may appear at line start if zh and loose assert_approx_equals: expected 69.66000366210938 +/- 1 but got 99.98999786376953
FAIL FE19  PRESENTATION FORM FOR VERTICAL HORIZONTAL ELLIPSIS may appear at line start if zh and loose assert_approx_equals: expected 69.66000366210938 +/- 1 but got 99.98999786376953
2024
文文文文文文․字字
文文文文文文
․字字
2025
文文文文文文‥字字
文文文文文文
‥字字
2026
文文文文文文…字字
文文文文文文
…字字
22EF
文文文文文文⋯字字
文文文文文文
⋯字字
FE19
文文文文文文︙字字
文文文文文文
︙字字
